implement test_astype
